move emitter details, cut autotuner changes
